[refactor] 잘못된 response DTO 패키지 위치 수정 (#18)
