+++ src/main/java/konkuk/chacall/domain/member/presentation/MemberController.java
@@ -6,9 +6,7 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import konkuk.chacall.domain.member.application.MemberService;
 import konkuk.chacall.domain.member.presentation.dto.request.*;
-import konkuk.chacall.domain.member.presentation.dto.response.ReservationForRatingResponse;
-import konkuk.chacall.domain.member.presentation.dto.response.SavedFoodTruckResponse;
-import konkuk.chacall.domain.member.presentation.dto.response.SavedFoodTruckStatusResponse;
+import konkuk.chacall.domain.member.presentation.dto.response.*;
 import konkuk.chacall.global.common.annotation.ExceptionDescription;
 import konkuk.chacall.global.common.annotation.UserId;
 import konkuk.chacall.global.common.dto.BaseResponse;
