+++ src/main/java/konkuk/chacall/domain/member/presentation/dto/response/MemberReservationHistoryResponse.java
@@ -1,4 +1,4 @@
-package konkuk.chacall.domain.member.presentation.dto.request;
+package konkuk.chacall.domain.member.presentation.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import konkuk.chacall.domain.foodtruck.domain.FoodTruck;
